Rename status module to _status.
We want its symbols imported from lsst.pipe.base, and this is a good way
to indicate that.
