+++ python/lsst/pipe/base/_status.py
